+++ src/Dnet.Blazor.Material/Components/FormField/Appearance.cs
@@ -1,9 +1,0 @@
-﻿namespace Dnet.Blazor.Material.Components.FormField;
-
-public enum Appearance
-{
-    Fill = 1,
-    Standard = 2,
-    Legacy = 3,
-    //Outline = 4
-}