+++ src/Dnet.Blazor.Material/Components/FormField/FloatLabel.cs
@@ -1,9 +1,0 @@
-﻿namespace Dnet.Blazor.Material.Components.FormField
-{
-    public enum FloatLabel
-    {
-        Always = 1,
-        AsUserTypes = 2,
-        Never = 3
-    }
-}