+++ src/Dnet.Blazor/Components/Dialog/Infrastructure/Models/DialogConfig.cs
@@ -10,6 +10,10 @@
 
         public string? PanelClass { get; set; }
 
+        public string? PanelStyle { get; set; }
+
+        public string? ThemeScope { get; set; }
+
         public bool HasBackdrop { get; set; } = true;
 
         public bool HasTransparentBackdrop { get; set; }