+++ src/Dnet.Blazor/Components/FloatingPanel/Infrastructure/Services/FloatingPanelService.cs
@@ -106,6 +106,9 @@ namespace Dnet.Blazor.Components.FloatingPanel.Infrastructure.Services
                 DisableBackdropClick  = floatingPanelConfig.DisableBackdropClick,
                 Width = floatingPanelConfig.Width == null ? "100vw" : floatingPanelConfig.Width + "px",
                 Height = floatingPanelConfig.Height == null ? "100vh" : floatingPanelConfig.Height + "px",
+                PanelClass = floatingPanelConfig.PanelClass,
+                PanelStyle = floatingPanelConfig.PanelStyle,
+                ThemeScope = floatingPanelConfig.ThemeScope,
                 GlobalPositionStrategy = globalPositionStrategy,
                 ComponentType = ComponentType.FloatingPanel
             };