+++ src/Dnet.Blazor/Components/Tooltip/Infrastructure/Models/TooltipConfig.cs
@@ -8,6 +8,10 @@
 
         public string? TooltipColor { get; set; }
 
+        public string? PanelStyle { get; set; }
+
+        public string? ThemeScope { get; set; }
+
         public string? Width { get; set; }
 
         public string? Height { get; set; }